feat: optimize project and rename to Omniview
Quality: clippy 2 errors + 34 warnings -> clean (never_loop drain helper,
mechanical --fix transforms); removed dead CameraUniform::new; added
scripts/gen-test-fixtures.sh to regenerate the /tmp test media.

Perf: audio sink takes &[f32] with a reused sample buffer (removes a
per-10ms ~3.8KB heap alloc); cpal callback bulk-drains the ring buffer in
one pass with a single volume pass; camera uniform dirty-flag skips the
per-frame write_buffer when nothing moved.

WSL DX: added README.md and Makefile (build/run/test/clippy/release/win);
'make test' auto-generates the test fixtures.

Release: [profile.release] thin LTO + codegen-units=1 + strip cuts the
binary 21M -> 14M (-33%); build-win.sh reports staged DLL count and exe
size. panic="abort" intentionally omitted: audio.rs relies on catch_unwind.

Rename: my-project -> omniview (Cargo package, tracing filter, window
title, example shim module, Windows exe name, README title).

+++ src/renderer/sphere.rs
@@ -62,8 +62,8 @@ impl Sphere {
                 let first = i * (sectors + 1) + j;
                 let second = first + sectors + 1;
                 indices.extend_from_slice(&[
-                    first as u32, second as u32, (first + 1) as u32,
-                    second as u32, (second + 1) as u32, (first + 1) as u32,
+                    first, second, (first + 1),
+                    second, (second + 1), (first + 1),
                 ]);
             }
         }
